+++ lib/mc-markdown/renderers/base.rb
@@ -1,11 +1,11 @@
 module MCMarkdown
   class Base < Redcarpet::Render::HTML
-    include ::MCMarkdown::Lists
-    include ::MCMarkdown::Image
-    include ::MCMarkdown::Links
-    include ::MCMarkdown::SafeMergeTags
-    include ::MCMarkdown::CommonMisspellings
-    include ::MCMarkdown::Blockquote
-    include ::MCMarkdown::Blocks
+    include ::MCMarkdown::Formatter::Lists
+    include ::MCMarkdown::Formatter::Image
+    include ::MCMarkdown::Formatter::Links
+    include ::MCMarkdown::Formatter::EscapeMergeTags
+    include ::MCMarkdown::Formatter::CommonMisspellings
+    include ::MCMarkdown::Formatter::Blockquote
+    include ::MCMarkdown::Formatter::Blocks
   end
 end