+++ spec/formatters/block_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe MCMarkdown::Blocks do
+describe MCMarkdown::Formatter::Blocks do
 
   it "generates a notes block" do
     test_fixture '/base/note'