+++ spec/formatters/blockquote_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe MCMarkdown::Blockquote do
+describe MCMarkdown::Formatter::Blockquote do
 
   it "splits blockquotes into two if there is a space between them" do
     test_fixture "base/blockquote"