+++ spec/formatters/common_misspellings_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe MCMarkdown::CommonMisspellings do
+describe MCMarkdown::Formatter::CommonMisspellings do
 
   it "filters out misspellings" do
     render_string("one&ldqou;&rdqou;&rsqou;").should == "<p>one&ldquo;&rdquo;&rsquo;</p>"